backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
asyncpg
